+++ Packages/Sources/TileContent/TileKit.Content.swift
@@ -1,0 +1,5 @@
+import TileCore
+
+public extension TileKit {
+    enum Content {}
+}