+++ Packages/Sources/TileKit/TileKit.Template.Rendering.swift
@@ -1,8 +1,0 @@
-public extension TileKit.Template {
-    protocol Rendering {
-        func render(
-            template: String,
-            context: Context,
-        ) throws -> String
-    }
-}