+++ Packages/Sources/TileKit/TileKit.Template.SimpleMustacheRendererError.swift
@@ -1,8 +1,0 @@
-public extension TileKit.Template {
-    enum SimpleMustacheRendererError: Error, Equatable {
-        case unterminatedTag(String)
-        case unexpectedClosingTag(String)
-        case missingValue(String)
-        case missingSectionEnd(String)
-    }
-}