+++ Packages/Sources/TileMarkdown/TileKit.Markdown.Rendering.swift
@@ -1,0 +1,9 @@
+import TileCore
+
+public extension TileKit.Markdown {
+    protocol Rendering {
+        func renderHTML(
+            _ markdown: String,
+        ) -> String
+    }
+}