+++ Packages/Sources/TileSource/TileKit.Source.FrontMatterParserError.swift
@@ -1,0 +1,8 @@
+import TileCore
+
+public extension TileKit.Source {
+    enum FrontMatterParserError: Error, Equatable {
+        case missingClosingSeparator
+        case invalidLine(String)
+    }
+}